+++ MatrixSDK/Crypto/Verification/Transactions/SAS/MXIncomingSASTransaction.h
@@ -22,12 +22,7 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  An handler on an incoming SAS device verification.
  */
-@interface MXIncomingSASTransaction : MXDefaultSASTransaction
-
-/**
- Accept the device verification request.
- */
-- (void)accept;
+@interface MXIncomingSASTransaction : MXLegacySASTransaction
 
 @end